Refactor start and stop services to single place
