+++ mgradm/cmd/start/podman.go
@@ -6,12 +6,9 @@ package start
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/uyuni-project/uyuni-tools/shared/podman"
+	"github.com/uyuni-project/uyuni-tools/mgradm/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
-	"github.com/uyuni-project/uyuni-tools/shared/utils"
 )
-
-var systemd podman.Systemd = podman.SystemdImpl{}
 
 func podmanStart(
 	_ *types.GlobalFlags,
@@ -19,10 +16,5 @@ func podmanStart(
 	_ *cobra.Command,
 	_ []string,
 ) error {
-	return utils.JoinErrors(
-		systemd.StartService(podman.DBService),
-		systemd.StartInstantiated(podman.ServerAttestationService),
-		systemd.StartInstantiated(podman.HubXmlrpcService),
-		systemd.StartService(podman.ServerService),
-	)
+	return podman.StartServices()
 }
