+++ mgradm/cmd/stop/podman.go
@@ -6,12 +6,9 @@ package stop
 
 import (
 	"github.com/spf13/cobra"
-	"github.com/uyuni-project/uyuni-tools/shared/podman"
+	"github.com/uyuni-project/uyuni-tools/mgradm/shared/podman"
 	"github.com/uyuni-project/uyuni-tools/shared/types"
-	"github.com/uyuni-project/uyuni-tools/shared/utils"
 )
-
-var systemd podman.Systemd = podman.SystemdImpl{}
 
 func podmanStop(
 	_ *types.GlobalFlags,
@@ -19,10 +16,5 @@ func podmanStop(
 	_ *cobra.Command,
 	_ []string,
 ) error {
-	return utils.JoinErrors(
-		systemd.StopInstantiated(podman.ServerAttestationService),
-		systemd.StopInstantiated(podman.HubXmlrpcService),
-		systemd.StopService(podman.ServerService),
-		systemd.StopService(podman.DBService),
-	)
+	return podman.StopServices()
 }
